refactor(shared): rename type files to match interface names
- Rename common.ts to IUnknownObject.ts
- Rename route.ts to IRoute.ts
- Update imports in action.wrapper.ts and routes.ts
